Module Core

Core language

The representation of the core language is not exposed. Instead, an abstract type is provided of both core language types and core language terms. Both terms and types support silent weakening: hence an element of tp or tm can be used in any scope.

type tp

type tm

val tp_of_tm : tm -> tp

Core language terms carry their own (unique) type.

Inspecting types

type tp_head = [

| `Pi

| `Sg

| `Bool

]

val tp_head : tp -> tp_head

The head of a type can be exposed in order to guide the elaborator. It is (surprisingly) unnecessary to expose any more data of a type to the elaborator.

Constructing well-typed terms

module Refiner : sig ... end

The refiner is the only way to construct terms. Any term constructed by the refiner is guaranteed to be well-typed, in the tradition of LCF.

Internals

module Private := Core__

The implementation of the core is non-trivial! It is worth studying the internals.