Module Core.Refiner

The refiner is the only way to construct terms. Any term constructed by the refiner is guaranteed to be well-typed, in the tradition of LCF.

Rule types

The refiner follows a version of the bidirectional typing discipline, dividing proofs into chk_rule and syn_rule. The purpose of the bidirectional division of labor is to enable many steps of refinement that would otherwise induce dozens of conversion checks to be collated in such a way that only one conversion check is required.

A related side-effect is that the refinement scripts contain very few annotations, drawing annotations inward from the goal and outward from the context.

type tp_rule

type chk_rule

type syn_rule

type tele_rule

Inference rules

val tl_nil : tele_rule

Telescopes

val tl_cons : string -> tp_rule -> (tm -> tele_rule) -> tele_rule

Booleans

val bool : tp_rule

val tt : chk_rule

val ff : chk_rule

Dependent product types

val pi : tp_rule -> (tm -> tp_rule) -> tp_rule

val lam : (tm -> chk_rule) -> chk_rule

val app : syn_rule -> chk_rule -> syn_rule

Dependent record types

val rcd_tp : tele_rule -> tp_rule

val rcd : chk_rule Basis.StringMap.t -> chk_rule

val proj : string -> syn_rule -> syn_rule

Dependent sum types

val sg : tp_rule -> (tm -> tp_rule) -> tp_rule

val pair : chk_rule -> chk_rule -> chk_rule

val fst : syn_rule -> syn_rule

val snd : syn_rule -> syn_rule

val chk_abort : chk_rule

Logical layer

Structural rules

val core : tm -> syn_rule

Every core language term carries has a unique type, and can hence be syn_rulethesized.

val conv : syn_rule -> chk_rule

The conversion rule appears in the bidirectional setting as the transition from synthesis to checking.

Rule combinators

val with_tp : (tp -> chk_rule) -> chk_rule

Failing rules

The following rules will fail with an exception.

val fail_tp : exn -> tp_rule

val fail_chk : exn -> chk_rule

val fail_syn : exn -> syn_rule