Module Effect.L

include Basis.Monad.S with type 'a m = 'a lm

type 'a m = 'a lm

val ret : 'a -> 'a m

val bind : 'a m -> ('a -> 'b m) -> 'b m

include Core__.EffectOps.L with type 'a m := 'a m

include Core__.EffectOps.G

type 'a m

include Basis.Error.Ops with type 'a m := 'a m

type 'a m

val throw : exn -> 'a m

Throw an exception to be captured within the monad.

val catch : 'a m -> (('a, exn) Stdlib.Result.t -> 'b m) -> 'b m

Handle an exception within the monad.

val theory : Core.Logic.thy m

Access the current logical theory.

val scope_thy : Core.Logic.update -> 'a m -> 'a m

Perform a monotone update to the current logical theory within a scope.

val env : [ `Tm of Core.Syntax.gtm | `Tp of Core.Syntax.gtp ] Core.Env.t m

Access the local variable environment

val bind_tm : Core.Syntax.gtp -> (Core.Syntax.gtm -> 'a m) -> 'a m

Bind a variable of a given type within a scope.

val append_tm : Core.Syntax.gtm -> 'a m -> 'a m

Append a term to the local environment within a scope.

val global : 'a gm -> 'a m

Forget the local variable environment.