Module Core__Effect.G

include Basis.Monad.S with type 'a m = 'a gm

type 'a m = 'a gm

val ret : 'a -> 'a m

val bind : 'a m -> ('a -> 'b m) -> 'b m

include Core__.EffectOps.G with type 'a m := 'a m

type 'a m

include Basis.Error.Ops with type 'a m := 'a m

type 'a m

val throw : exn -> 'a m

Throw an exception to be captured within the monad.

val catch : 'a m -> (('a, exn) Stdlib.Result.t -> 'b m) -> 'b m

Handle an exception within the monad.

val theory : Core.Logic.thy m

Access the current logical theory.

val scope_thy : Core.Logic.update -> 'a m -> 'a m

Perform a monotone update to the current logical theory within a scope.

val local : Core.Syntax.env -> 'a lm -> 'a m

Execute with a local variable environment.