Module Frontend

The source language

We begin by defining a naive source language.

type code =

| R of rcode

| L of lcode

The central idea to the elaboration algorithm is to distinguish introduction forms from elimination forms; unlike some classic bidirectional algorithms, this distinction does not line up exactly with checking vs. synthesis, but it interacts with it in a non-trivial way: we only syn_rulethesize elimination forms at positive types.

and rcode =

| Bool

| Pi of string * code * code

| Sg of string * code * code

| Tt

| Ff

| Lam of string * code

| Pair of code * code

rcode is a type of introduction forms

and lcode =

| Var of string

| App of code * code

| Fst of code

| Snd of code

| Core of Core.tm

lcode is a type of elimination forms. Included via Core is the collection of all core-language terms; this embedding is used to crucial effect by the elaborator.

Elaboration

module R = Core.Refiner

module Elaborator : sig ... end

Distillation

module Distiller : sig ... end

The distiller takes a core-language term and turns it into a source language code.