Module Frontend.Elaborator

type resolver

include Basis.Reader.S with type local = resolver

include Basis.Monad.Trans

type 'a n

include Basis.Monad.S

type 'a m

val ret : 'a -> 'a m

val bind : 'a m -> ('a -> 'b m) -> 'b m

val lift : 'a n -> 'a m

include Basis.Reader.Ops with type 'a m := 'a m

type 'a m

type local

val read : local m

val locally : (local -> local) -> 'a m -> 'a m

val read : local m

val locally : (local -> local) -> 'a m -> 'a m

val reader : (local -> 'a n) -> 'a m

val run : local -> 'a m -> 'a n

val elab_chk_code : code -> R.chk_rule m

The main entry-point: check a piece of code against a type.

val elab_chk_rcode : rcode -> R.chk_rule m

Checking introduction forms against their types.

val elab_chk_lcode : lcode -> R.chk_rule m

Rather than transitioning immediately to synthesize when we hit an lcode, we perform type-directed eta expansion. This is the main ingredient to enable smooth elaboration of subtypes, including the "retyping principles" familiar from ML modules.

val elab_syn_lcode : lcode -> R.syn_rule m

Elaborating an elimination form.

val elab_tp_code : code -> R.tp_rule m

Elaborate a type

val elab_tp_rcode : rcode -> R.tp_rule m