Module Frontend__Code

type code =

| R of rcode

| L of lcode

The central idea to the elaboration algorithm is to distinguish introduction forms from elimination forms; unlike some classic bidirectional algorithms, this distinction does not line up exactly with checking vs. synthesis, but it interacts with it in a non-trivial way: we only synthesize elimination forms at positive types.

and rcode =

| Bool

| Pi of string * code * code

| RcdTp of tele_code

| Sg of string * code * code

| Tt

| Ff

| Lam of string * code

| Pair of code * code

| Rcd of code Basis.StringMap.t

rcode is a type of introduction forms

and lcode =

| Var of string

| App of code * code

| Fst of code

| Snd of code

| Proj of string * code

| Core of Core.tm

lcode is a type of elimination forms. Included via Core is the collection of all core-language terms; this embedding is used to crucial effect by the elaborator.

and tele_code =

| TlNil

| TlCons of string * code * tele_code