Module Core.Refiner

The refiner is the only way to construct terms. Any term constructed by the refiner is guaranteed to be well-typed, in the tradition of LCF.

Rule types

The refiner follows a version of the bidirectional typing discipline, dividing proofs into chk and syn. The purpose of the bidirectional division of labor is to enable many steps of refinement that would otherwise induce dozens of conversion checks to be collated in such a way that only one conversion check is required.

A related side-effect is that the refinement scripts contain very few annotations, drawing annotations inward from the goal and outward from the context.

type chk

type syn

Runners

val run_syn : syn -> tm

A syn refinement script can be executed to yield an abstract term; when it returns a value, that value is guaranteed to be well-typed.

val run_chk : chk -> tp -> tm

A chk refinement script may be executed relative to a given type; when it returns a value, that value is guaranteed to have the type given.

val chk_to_ltm : chk -> tp -> Syntax.ltm

It is also useful to read a local term off a refinement script.

Inference rules

Booleans

val tt : chk

val ff : chk

Dependent product types

val lam : (tm -> chk) -> chk

val app : syn -> chk -> syn

Dependent sum types

val pair : chk -> chk -> chk

val fst : syn -> syn

val snd : syn -> syn

Structural rules

val core : tm -> syn

Every core language term carries has a unique type, and can hence be synthesized.

val conv : syn -> chk

The conversion rule appears in the bidirectional setting as the transition from synthesis to checking.

Rule combinators

val with_tp : (tp -> chk) -> chk