Module Core__.Syntax

Core language representations

The core language syntax is split into two representations, after Coquand: a local form that is sensitive to the context (using De Bruijn indices), and a global form that is insensitive to the context (using De Bruijn levels).

The "local" syntax corresponds to ordinary syntax, and the "global" syntax corresponds to weak head normal forms in many NbE-style implementations. We make no attempt to restrict local syntax to normal forms; unlike conventional implementation, we freely interleave the local and the global syntax when it saves us some computations (see the annotations on LLam and LRcd for instance).

Logical layer

type 'v prop =

| PVar of 'v

| PTop

| PBot

type gprop = Core.Env.lvl prop

type lprop = Core.Env.ix prop

Representation of types

type ltp =

| LPi of ltp * ltp

| LRcdTp of string list * ltele

| LBool

| LAbortTp

| LTpVar of Core.Env.ix

and gtp =

| GPi of gfam

| GRcdTp of string list * gtele

| GBool

| GAbortTp

and gfam = gtp * ltp * env

and gtele =

| GTlNil

| GTlCons of gtp * ltele * env

and ltele =

| LTlNil

| LTlCons of ltp * ltele

Representation of terms

and ltm =

| LVar of Core.Env.ix

| LTt

| LFf

| LLam of gfam * ltm

| LApp of ltm * ltm

| LRcd of string list * gtele * ltm Basis.StringMap.t

| LProj of string * ltm

| LAbort

and gtm =

| GTt

| GFf

| GLam of gfam * ltm * env

| GRcd of string list * gtele * gtm Basis.StringMap.t

| Glued of (gneu, ltm) glued

| GAbort

and gneu =

| GVar of Core.Env.lvl

| GSnoc of gneu * gfrm

and gfrm =

| GProj of string

| GApp of gtm

and ('b, 'a) glued =

| Gl of {

supp : gprop;

gtp : gtp;

base : 'b;

part : 'a;

env : env;

}

and 'a part =

| Prt of {

supp : gprop;

part : 'a;

env : env;

}

and env = cell Core.Env.t

and cell = [

| `Tm of gtm

| `Tp of gtp

| `Prop of gprop

]

Convenience

type tp_head = [

| `Pi

| `Rcd of string list

| `Bool

| `Abort

]

val tp_head : gtp -> tp_head

Project the name of the head constructor of a type; useful for guiding elaboration.

val tp_of_gtm : gtm -> gtp

Project the type of a term: this is efficient and non-recursive.

Glued terms

val glued_to_part : ('b, 'a) glued -> 'a part

Project the partial element from a glued term.

val stable_glued : gtp -> 'b -> ('b, ltm) glued

Construct a stable glued term, i.e. one form whom the base is nowhere unstable.

Restricting to partial elements

val gtm_to_part : gprop -> gtm -> ltm part

Restrict a total term to a partial term.

val gtp_to_part : gprop -> gtp -> ltp part

Restrict a total type to a partial type.

Projecting boundaries

The following functions project the partial (terms, types) that a (term, type) must compute to; when the input has a stable head constructor, the empty partial element is returned.

val gtm_bdry : gtm -> ltm part

val gtp_bdry : gtp -> ltp part